requirements/local.txt
-r base.txt
django-debug-toolbar==1.7
django-debug-toolbar-template-timings==0.7
mock==2.0.0
psycopg2==2.7.1
model_mommy==1.3.2